tests/ipc/pyproject.toml
[tool.poetry]
name = "ggtest"
version = "0.1.0"
description = "Tests for greengrass lite IPC"
authors = ["Amazon.com, Inc."]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.3"
awsiotsdk = "^1.19.0"
boto3 = "^1.34.2"
paho-mqtt = "^1.6.1"
pyyaml = "^6.0.1"
pytest-json-report = "^1.5.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
pyright = "^1.1.341"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.pytest.ini_options]
# adopts = [
#     "--strict-markets",
#     "--strict-config",
#     "-ra"
# ]
# testpaths = "tests"
# markers = [

# ]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
